+++ test/user.test.js
@@ -1,5 +1,4 @@
 const userService = require("../service/user.service");
-// const userValidation = require("../models/userValidation");
 
 describe("login한 user", () => {
   const user = {
